requirements.txt
diffusers==0.21.0
evaluate==0.4.0
keras==2.13.1
langchain==0.0.288
langchain_experimental==0.0.17
matplotlib==3.7.3
numpy==1.22.4
opencv_contrib_python==4.8.0.76
opencv_python==4.8.0.76
opencv_python_headless==4.8.0.76
pandas==2.1.0
Pillow==10.0.0
Requests==2.31.0
tensorflow==2.13.0
tensorflow_datasets==4.8.2
torch==2.0.1
torchsummary==1.5.1
torchtext==0.15.2
torchvision==0.15.2
transformers==4.33.1
ultralytics==8.0.178
